fix: 자동차 이름 관련 비즈니스 규칙을 Car 생성자에서 체크하도록 이동
