+++ src/main/java/racing/domain/Car.java
@@ -1,20 +1,25 @@
 package racing.domain;
 
 import racing.Constant;
-import racing.utils.ValidationUtil;
 
 public class Car {
     private String name;
     private int position;
 
     public Car(final String name) {
-        ValidationUtil.validateName(name);
+        if (name == null || name.isBlank()) {
+            throw new IllegalArgumentException(Constant.CAR_NAME_BLANK_VALIDATION_ERROR);
+        }
+
+        if (name.length() > 5) {
+            throw new IllegalArgumentException(Constant.CAR_NAME_LENGTH_VALIDATION_ERROR);
+        }
+
         this.name = name;
         this.position = Constant.START_POSITION;
     }
 
     public Car(final String name, int position) {
-        ValidationUtil.validateName(name);
         this.name = name;
         this.position = position;
     }
