+++ src/test/java/racing/domain/CarTest.java
@@ -2,8 +2,10 @@ package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.Constant;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class CarTest {
 
@@ -29,6 +31,22 @@ class CarTest {
         Car car = new Car("pobi", 5);
         assertThat(car.max(3)).isEqualTo(5);
         assertThat(car.isMatch(5));
+    }
+
+    @Test
+    @DisplayName("차 이름을 입력하지 않을 경우 RuntimeException 예외가 발생해야 한다")
+    void validate_carName_null() {
+        assertThatThrownBy(() -> new Car(null))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining(Constant.CAR_NAME_BLANK_VALIDATION_ERROR);
+    }
+
+    @Test
+    @DisplayName("5자 이상의 이름을 입력할 경우 RuntimeException 예외가 발생해야 한다")
+    void validate_carName_length() {
+        assertThatThrownBy(() -> new Car("HarryPorter"))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining(Constant.CAR_NAME_LENGTH_VALIDATION_ERROR);
     }
 
 }
