+++ src/test/java/racing/utils/ValidationUtilTest.java
@@ -6,7 +6,6 @@ import racing.Constant;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static racing.utils.ValidationUtil.validateInput;
-import static racing.utils.ValidationUtil.validateName;
 
 class ValidationUtilTest {
 
@@ -34,19 +33,4 @@ class ValidationUtilTest {
                 .hasMessageContaining(Constant.CAR_NUMBER_VALIDATION_ERROR);
     }
 
-    @Test
-    @DisplayName("차 이름을 입력하지 않을 경우 RuntimeException 예외가 발생해야 한다")
-    void validate_carName_null() {
-        assertThatThrownBy(() -> validateName(null))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining(Constant.CAR_NAME_BLANK_VALIDATION_ERROR);
-    }
-
-    @Test
-    @DisplayName("5자 이상의 이름을 입력할 경우 RuntimeException 예외가 발생해야 한다")
-    void validate_carName_length() {
-        assertThatThrownBy(() -> validateName("HarryPorter"))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining(Constant.CAR_NAME_LENGTH_VALIDATION_ERROR);
-    }
 }
